refactor filter controls interface
