+++ aries-site/src/examples/templates/FilterControls/FilterControls.js
@@ -4,12 +4,21 @@ import { Box } from 'grommet';
 
 import { Filters, ResultsSummary, SearchFilter, useFilters } from '.';
 
-export const FilterControls = ({ data }) => {
-  const { setData } = useFilters();
+export const FilterControls = ({ data, filters }) => {
+  const { setData, setFilterAttributes, syncFilteredResults } = useFilters();
 
   useEffect(() => {
     setData(data);
   }, [data, setData]);
+
+  useEffect(() => {
+    setFilterAttributes(filters);
+  }, [filters, setFilterAttributes]);
+
+  /* keep filteredResults up to date as parent data set may change */
+  useEffect(() => {
+    syncFilteredResults();
+  }, [syncFilteredResults]);
 
   return (
     <Box gap="xsmall">
@@ -24,4 +33,12 @@ export const FilterControls = ({ data }) => {
 
 FilterControls.propTypes = {
   data: PropTypes.array,
+  filters: PropTypes.arrayOf(
+    PropTypes.shape({
+      property: PropTypes.string,
+      label: PropTypes.string,
+      filterType: PropTypes.oneOf(['CheckBoxGroup', 'RangeSelector']),
+      inputProps: PropTypes.object,
+    }),
+  ),
 };
